+++ frontend/src/components/molecules/ProjectViews/ViewsDropdown.vue
@@ -96,7 +96,8 @@ const images = {
     WhiteboardView: {image: require('@/assets/images/png/Board.png'), description: 'whiteboard_view'},
     CanvasView: {image: require('@/assets/images/png/Activity.png'), description: 'canvas_view'},
     MapView: {image: require('@/assets/images/png/Workload.png'), description: 'map_view'},
-    ProjectDashboard: {image: require('@/assets/images/png/Workload.png'), description: 'dashboard_view'}
+    ProjectDashboard: {image: require('@/assets/images/png/Workload.png'), description: 'dashboard_view'},
+    DocsView: {image: require('@/assets/images/png/Activity.png'), description: 'docs_view'}
 }
 const viewItem = ref('')
 const companyId = inject('$companyId')